Moved and refactored networking logic for TcpClientUser
